Move most Drivetrain-related classes into their own package
